[C] Migrate resource-preview
