+++ client/src/frontend/components/resource-preview/Types/Image.js
