+++ client/src/frontend/components/resource-preview/Types/Video.js
