+++ client/src/frontend/components/resource-preview/Types/index.js
